refactor nurse activity - added nurse fragment
