+++ app/src/main/java/pe/edu/upc/attentionapp/controllers/fragments/HomeFragment.kt
@@ -3,4 +3,5 @@ package pe.edu.upc.attentionapp.controllers.fragments
 import androidx.fragment.app.Fragment
 
 class HomeFragment: Fragment() {
+
 }
