+++ app/src/main/res/layout/fragment_nurses.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".controllers.activities.NursesAvailableActivity">
+    tools:context=".controllers.fragments.NursesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/nursesRecyclerView"
