+++ app/src/main/res/values/strings.xml
@@ -19,8 +19,8 @@
 
     <!--Home-->
     <string name="title_home">Home</string>
-    <string name="title_sources">Sources</string>
     <string name="title_settings">Settings</string>
     <string name="tile_favorites">Favorites</string>
+    <string name="title_nurses">Enfermeras</string>
 
 </resources>
